move AutonomousUseCase into base package too
- now fits better semantically (esp. as soon as we split up agents
  from use-cases)
